+++ app/[locale]/components/button/button.tsx
@@ -4,12 +4,13 @@ import './button.css';
 type ButtonType = {
   title:string,
   type?:"submit" | "reset" | "button" | undefined,
-  disabled?:boolean
+  disabled?:boolean,
+  bigger?:boolean
 }
 
-const Button = ({title, type, disabled}:ButtonType) =>{
+const Button = ({title, type, disabled, bigger}:ButtonType) =>{
   return(
-    <button type={type} disabled={disabled} className='button-gradient'><span className='span-button'>{title}</span></button>
+    <button type={type} disabled={disabled} className='button-gradient'><span className={`span-button ${bigger && 'bigger'}`}>{title}</span></button>
   )
 }
