+++ app/[locale]/data/faqs.ts
@@ -1,8 +1,15 @@
 export const faqs = [
-  {question:'Care Model Behavior', answer:"There is no doubt that healthcare’s multifaceted structure prompts challenges throughout the care continuum. Meanwhile, patients increasingly expect the convenience of a seamless, quality healthcare experience from clinic to the hospital, to post-acute care environments, and ultimately to their home. So, healthcare industry partners must work together to determine how to best employ technology to improve the patient's experience, accelerate care delivery—no matter where it happens—and lower costs—all at the same time. Novel care models have the potential to deliver integrated patient care across the continuum that results in positive quality outcomes and an outstanding experience for all involved."},
-  {question:'Care Model Behavior', answer:"There is no doubt that healthcare’s multifaceted structure prompts challenges throughout the care continuum. Meanwhile, patients increasingly expect the convenience of a seamless, quality healthcare experience from clinic to the hospital, to post-acute care environments, and ultimately to their home. So, healthcare industry partners must work together to determine how to best employ technology to improve the patient's experience, accelerate care delivery—no matter where it happens—and lower costs—all at the same time. Novel care models have the potential to deliver integrated patient care across the continuum that results in positive quality outcomes and an outstanding experience for all involved."},
-  {question:'Care Model Behavior', answer:"There is no doubt that healthcare’s multifaceted structure prompts challenges throughout the care continuum. Meanwhile, patients increasingly expect the convenience of a seamless, quality healthcare experience from clinic to the hospital, to post-acute care environments, and ultimately to their home. So, healthcare industry partners must work together to determine how to best employ technology to improve the patient's experience, accelerate care delivery—no matter where it happens—and lower costs—all at the same time. Novel care models have the potential to deliver integrated patient care across the continuum that results in positive quality outcomes and an outstanding experience for all involved."},
-  {question:'Care Model Behavior', answer:"There is no doubt that healthcare’s multifaceted structure prompts challenges throughout the care continuum. Meanwhile, patients increasingly expect the convenience of a seamless, quality healthcare experience from clinic to the hospital, to post-acute care environments, and ultimately to their home. So, healthcare industry partners must work together to determine how to best employ technology to improve the patient's experience, accelerate care delivery—no matter where it happens—and lower costs—all at the same time. Novel care models have the potential to deliver integrated patient care across the continuum that results in positive quality outcomes and an outstanding experience for all involved."},
-  {question:'Care Model Behavior', answer:"There is no doubt that healthcare’s multifaceted structure prompts challenges throughout the care continuum. Meanwhile, patients increasingly expect the convenience of a seamless, quality healthcare experience from clinic to the hospital, to post-acute care environments, and ultimately to their home. So, healthcare industry partners must work together to determine how to best employ technology to improve the patient's experience, accelerate care delivery—no matter where it happens—and lower costs—all at the same time. Novel care models have the potential to deliver integrated patient care across the continuum that results in positive quality outcomes and an outstanding experience for all involved."},
-  {question:'Care Model Behavior', answer:"There is no doubt that healthcare’s multifaceted structure prompts challenges throughout the care continuum. Meanwhile, patients increasingly expect the convenience of a seamless, quality healthcare experience from clinic to the hospital, to post-acute care environments, and ultimately to their home. So, healthcare industry partners must work together to determine how to best employ technology to improve the patient's experience, accelerate care delivery—no matter where it happens—and lower costs—all at the same time. Novel care models have the potential to deliver integrated patient care across the continuum that results in positive quality outcomes and an outstanding experience for all involved."},
+  {question:'public', answer:"public"},
+  {question:'topics', answer:"topics"},
+  {question:'register', answer:"register"},
+  {question:'ticket', answer:"ticket"},
+  {question:'discount', answer:"discount"},
+  {question:'accreditation', answer:"accreditation"},
+  {question:'networking', answer:"networking"},
+  {question:'sponsor', answer:"sponsor"},
+  {question:'benefits', answer:"benefits"},
+  {question:'levels', answer:"levels"},
+  {question:'exhibition', answer:"exhibition"},
+  {question:'event', answer:"event"},
+  {question:'exclusiveness', answer:"exclusiveness"},
 ]
