+++ app/[locale]/homeComponents/sponsorsContact.css
@@ -1,73 +1,34 @@
-.form-cont {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  background-color: black;
-}
-
-.form-cont> button {
-  width: 7rem;
-}
-
-.input-form {
-  margin: 20px 25px;
-  width: 200px;
-  display: block;
-  border: none;
-  padding-top: 20px;
-  padding-bottom: 5px;
-  border-bottom: solid 1px white;
-  transition: all 0.3s cubic-bezier(.64,.09,.08,1);
-  background: linear-gradient(to bottom, rgba(255,255,255,0) 96%, white 4%);
-  background-position: -200px 0;
-  background-size: 200px 100%;
+.cont-form {
+  background-image: url('/backg-contact.png');
+  background-position: center;
   background-repeat: no-repeat;
-  color: white;
+  background-size: cover;
+  display: flex;
+  justify-content: space-evenly;
 }
 
-.textarea {
-  resize: none;
-
+.contact-text{
+  width: 30%;
 }
 
-.input-form:focus, .input-form:valid {
-  box-shadow: none;
-  outline: none;
-  background-position: 0 0;
+.contact-text>h4 {
+  font-size: 2rem;
 }
 
-.input-form::-webkit-input-placeholder{
-  color: white;
-  font-size: 13px;
-  transform: translateY(-20px);
-  visibility: visible;
+.contact-text>p{
+  font-size: 1.5rem;
 }
 
-.captcha-style {
-  margin: 1rem;
-}
+@media screen and (max-width:800px) {
+  .cont-form {
+    flex-direction: column-reverse;
+    background-position: left;
+    align-items: center;
+  }
 
-.error-message{
-  font-size: 15px;
-  color: rgb(96, 0, 86);
-}
-
-.popover {
-  background-color: green;
-  width: 80%;
-  border-radius: 10px;
-  border: 2px solid white;
-  padding: 0.5rem 0;
-  text-align: center;
-  margin-top: 1rem;
-}
-
-.popover > p {
-  color: white;
-}
-
-@media screen and (max-width:900px) {
-  .form-cont {
-    margin-bottom: 2rem;
+  .contact-text{
+    width: 80%;
+    text-align: center;
+    margin-bottom: 3rem;
   }
 }
